requirements.txt
asgiref==3.2.7
certifi==2020.4.5.1
chardet==3.0.4
Django==3.0.6
django-oauth-toolkit==1.3.2
djangorestframework==3.11.0
factory-boy==2.12.0
Faker==4.1.0
idna==2.9
oauthlib==3.1.0
psycopg2==2.8.5
Pygments==2.6.1
python-dateutil==2.8.1
python-dotenv==0.13.0
pytz==2020.1
requests==2.23.0
six==1.15.0
sqlparse==0.3.1
text-unidecode==1.3
urllib3==1.25.9
